1494 - Move percentage share function to expressions utils
